Moved eibc middleware logic to a seperate module.
